+++ src/main/java/roomescape/auth/JwtTokenProvider.java
@@ -9,8 +9,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
-import roomescape.member.AuthenticatedMember;
-import roomescape.member.Role;
+import roomescape.member.domain.AuthenticatedMember;
+import roomescape.member.domain.Role;
 
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
