+++ src/main/java/roomescape/login/LoginService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
-import roomescape.member.AuthenticatedMember;
-import roomescape.member.Member;
-import roomescape.member.infrastructure.MemberRepository;
+import roomescape.member.domain.AuthenticatedMember;
+import roomescape.member.domain.Member;
+import roomescape.member.application.port.out.MemberRepository;
 
 @Service
 @RequiredArgsConstructor
