+++ src/main/java/roomescape/member/domain/Role.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.domain;
 
 public enum Role {
     USER,
