+++ src/main/java/roomescape/reservation/adapter/in/web/UserReservationController.java
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.common.api.ApiResponse;
-import roomescape.member.AuthenticatedMember;
-import roomescape.member.LoginMember;
+import roomescape.member.domain.AuthenticatedMember;
+import roomescape.member.domain.LoginMember;
 import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.application.dto.request.ReservationSaveRequest;
 import roomescape.reservation.application.dto.response.ReservationDetailFindResponse;
