+++ src/main/java/roomescape/waiting/adapter/in/web/WaitingController.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.common.api.ApiResponse;
-import roomescape.member.AuthenticatedMember;
-import roomescape.member.LoginMember;
+import roomescape.member.domain.AuthenticatedMember;
+import roomescape.member.domain.LoginMember;
 import roomescape.waiting.application.WaitingService;
 import roomescape.waiting.application.dto.request.WaitingRequest;
 import roomescape.waiting.application.dto.response.WaitingResponse;
