+++ src/test/java/roomescape/reservation/ReservationControllerTest.java
@@ -14,8 +14,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import roomescape.common.api.ApiResponse;
-import roomescape.member.AuthenticatedMember;
-import roomescape.member.Role;
+import roomescape.member.domain.AuthenticatedMember;
+import roomescape.member.domain.Role;
 import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.application.dto.request.ReservationSaveRequest;
 import roomescape.reservation.application.dto.response.ReservationDetailFindResponse;
