+++ src/test/java/roomescape/waiting/WaitingControllerTest.java
@@ -13,8 +13,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import roomescape.common.api.ApiResponse;
-import roomescape.member.AuthenticatedMember;
-import roomescape.member.Role;
+import roomescape.member.domain.AuthenticatedMember;
+import roomescape.member.domain.Role;
 import roomescape.waiting.application.WaitingService;
 import roomescape.waiting.application.dto.request.WaitingRequest;
 import roomescape.waiting.application.dto.response.WaitingResponse;
